+++ src/as_debug/display_debug.rs
@@ -1,14 +1,12 @@
 use core::fmt::Display;
 
-use crate::as_debug::{AsDisplayDebug, DisplayDebugged};
+use crate::as_debug::{DisplayAsDebug, DisplayDebugged};
 
 /// A trait to convert a type to use its [`Display`] implementation for [`Debug`].
 pub trait DisplayDebug: Display {
     /// Wraps a borrowed value in an adaptor that enable the values [`Display`] implementation to be
     /// used for its [`Debug`] implementation.
     ///
-    /// See [`AsDisplayDebug`] for more information.
-    ///
     /// # Examples
     ///
     /// ```rust
@@ -16,20 +14,17 @@ pub trait DisplayDebug: Display {
     /// use std::net::IpAddr;
     ///
     /// let ip = IpAddr::V4("127.0.0.1".parse().unwrap());
-    /// assert_eq!(format!("{:?}", ip.as_debug()), "127.0.0.1");
-    ///
-    /// // borrowed value's display implementation is still used for display
-    /// assert_eq!(format!("{}", ip.as_debug()), "127.0.0.1");
+    /// assert_eq!(format!("{:?}", ip.display_as_debug()), "127.0.0.1", "display used for debug");
+    /// assert_eq!(format!("{}", ip.display_as_debug()), "127.0.0.1", "display unchanged");
     /// ```
-    fn as_debug(&'_ self) -> AsDisplayDebug<'_, Self> {
-        AsDisplayDebug(self)
+    fn display_as_debug(&'_ self) -> DisplayAsDebug<'_, Self> {
+        DisplayAsDebug(self)
     }
 
     /// Wraps a owned value in an adaptor that enable the values [`Display`] implementation to be
     /// used for its [`Debug`] implementation.
     ///
-    /// Unless ownership is not necessary, favor [`DisplayDebug::as_debug`].
-    /// See [`DisplayDebugged`] for more information.
+    /// Unless ownership is not necessary, favor [`DisplayDebug::display_as_debug`].
     ///
     /// # Examples
     ///
@@ -38,12 +33,10 @@ pub trait DisplayDebug: Display {
     /// use std::net::IpAddr;
     ///
     /// let ip = IpAddr::V4("127.0.0.1".parse().unwrap());
-    /// assert_eq!(format!("{:?}", ip.to_debug()), "127.0.0.1");
-    ///
-    /// // owned value's display implementation is still used for display
-    /// assert_eq!(format!("{}", ip.to_debug()), "127.0.0.1");
+    /// assert_eq!(format!("{:?}", ip.wrap_display_as_debug()), "127.0.0.1", "display used for debug");
+    /// assert_eq!(format!("{}", ip.wrap_display_as_debug()), "127.0.0.1", "display unchanged");
     /// ```
-    fn to_debug(self) -> DisplayDebugged<Self>
+    fn wrap_display_as_debug(self) -> DisplayDebugged<Self>
     where
         Self: Sized,
     {