+++ src/option/opaque_dbg.rs
@@ -1,4 +1,6 @@
-use core::fmt::Debug;
+use core::fmt::{Debug, Formatter};
+
+use crate::Opaque;
 
 /// A [`Option<T>`] wrapper that implements [`Debug`] with opaque Some values.
 ///
@@ -7,9 +9,9 @@ use core::fmt::Debug;
 pub struct OpaqueOptionDebug<'a, T>(pub &'a Option<T>);
 
 impl<T> Debug for OpaqueOptionDebug<'_, T> {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         match self.0 {
-            Some(_) => f.write_str("Some(..)"),
+            Some(_) => f.debug_tuple("Some").field(&Opaque).finish(),
             None => f.write_str("None"),
         }
     }