+++ tests/opaque.rs
@@ -1,0 +1,7 @@
+mod common;
+
+use display_as_debug::Opaque;
+
+use crate::common::*;
+
+test_fmt!(opaque_debug, Opaque, "{:?}", "..");